+++ libs/skill-pronote/src/lib/shared/kid.ts
@@ -1,0 +1,10 @@
+import { GenericInterface } from "./types";
+
+export interface Kid extends GenericInterface {
+	id: string,
+	firstName: string,
+	lastName: string,
+	class: string,
+	dateOfBirth: Date,
+	ref?: any
+}